Added YouTube Tutorial Pop-up and Minor Fix
Add the YouTube Pop-up box when the user clicks on "How to use?". Clicking on the cross button now also supports closing the video completely whereas earlier it used to close only the modal letting the video play in the background. It's fixed now.

In the footer, replaced Electron with Github student dev pack and Electron didn't actually contribute to this web version of GS. But GSDP surely did by providing the free domain for the first year :+1: